Okay sort of works
